Changed howteleport() and battleChance() work. Changed how battles work.
Originally, we were teleporting to another map for our battles (or our battle placeholder). This was causing problems with the teleport function and was generally a pain. Ideally, we don't want the battles to take place in a map, unless we want to do something in real time. I always envisioned a menu system for the battle, and as a result, I've removed the battle map and the teleportation. Now, we have 2 new functions: renderBattle() and endBattle(). renderBattle() draws a placeholder image over the map and sets a boolean that stops the player from moving behind the scenes. This is where we should add code to make the battle do what we want it to do. endBattle resets the boolean to allow the player to move again, and redraws the map. This is what we should call when whatever conditions we want to set are met in the battle. For now, that condition is waiting 2 seconds (with the setTimeout() function).