Fix bugs and add functionality to Teleportation
Fix some bugs in the random enemy creation. Set enemy health in the switch
used to choose an enemy rather than a separate function.
Set music for the overworld and battle to loop, and set battle music to
restart for each battle.
Change testAnimation() to teleportAnimation and set it to be called during
teleportation at the portal. Add the hero sprite on top of the teleportation
animation to make it more clear that the player is moving. Change the
map in the middle of the animation rather than just as it begins.